frontend/Login: Revamped Login page with the app's design system and moved it to /Pages
