+++ frontend/src/Components/Buttons/Buttons.css
@@ -1,8 +1,10 @@
 .button {
-    all: unset;
+    all   : unset;
+    cursor: pointer;
 }
 
 .button-flat {
+    all             : unset;
     width           : 100%;
     border          : none;
     border-radius   : 0.4rem;
@@ -19,6 +21,7 @@
 }
 
 .button-raised {
+    all             : unset;
     width           : 100%;
     border-radius   : 0.6rem;
     padding         : .8rem 1rem;
