+++ frontend/src/Components/Buttons/Buttons.js
@@ -5,18 +5,18 @@ import './Buttons.css';
 
 //Button returns a button with three possible types: flat, raised or round
 export default function Button({
-    text, link, type, icon, onClick, tooltip
+    btnVariant, text, link, type, icon, onClick, tooltip
 }) {
     //link, text, type, icon, onClick and tooltip are passed through props to this component
-    if ((typeof link === 'string' && typeof type === 'string')) {
+    if ((typeof link === 'string' && typeof btnVariant === 'string')) {
         //Makes sure link passed is a valid string
         if (icon && typeof icon === 'string') {
             //Checks if an icon was passed and is a valid string
-            if (type === 'flat-icon') {
+            if (btnVariant === 'flat-icon') {
                 if (link.startsWith('/') || link === '') {
                     //For local or empty links, a react-router link as a button is returned with its provided icon
                     return (
-                        <button className='button' onClick={onClick}>
+                        <button className='button' onClick={onClick} type={type}>
                             <Link className="button-flat icon" to={link}>
                                 {text ? <span>{text}</span> : null}
                                 <span className="iconify" data-icon={icon}></span>
@@ -33,10 +33,10 @@ export default function Button({
                     )
                 }
                 //Same checking for the raised with icon type button
-            } else if (type === 'raised-icon') {
+            } else if (btnVariant === 'raised-icon') {
                 if (link.startsWith('/') || link === '') {
                     return (
-                        <button className='button' onClick={onClick}>
+                        <button className='button' onClick={onClick} type={type}>
                             <Link className="button-raised icon" to={link}>
                                 {text ? <span>{text}</span> : null}
                                 <Icon icon={icon} />
@@ -51,10 +51,10 @@ export default function Button({
                         </a>
                     )
                 }
-            } else if (type === 'round') {
+            } else if (btnVariant === 'round') {
                 //For a round icon, only a react-router link with an icon is returned. No text included.
                 return (
-                    <button className='button' onClick={onClick}>
+                    <button className='button' onClick={onClick} type={type}>
                         <Link className="button-round tooltip" to={link}>
                             <Icon icon={icon} />
                             {(tooltip && tooltip !== '')
@@ -65,15 +65,15 @@ export default function Button({
                 )
                 //An icon was passed as prop but button name is not correct type
             } else {
-                console.log('Wrong button type was passed to Button component');
+                console.error('Wrong button name');
                 return null;
             }
             //Same functionality and ckeckings as with icon-included buttons but for only text buttons
         } else {
-            if (type === 'flat') {
+            if (btnVariant === 'flat') {
                 if (link.startsWith('#')) {
                     return (
-                        <button className='button' onClick={onClick}>
+                        <button className='button' onClick={onClick} type={type}>
                             <a className="button-flat" href={link}>
                                 <span>{text}</span>
                             </a>
@@ -81,17 +81,17 @@ export default function Button({
                     )
                 } else {
                     return (
-                        <button className='button' onClick={onClick}>
+                        <button className='button' onClick={onClick} type={type}>
                             <Link className="button-flat" to={link}>
                                 <span>{text}</span>
                             </Link>
                         </button>
                     )
                 }
-            } else if (type === 'raised') {
-                if (link.startsWith('#')) {
+            } else if (btnVariant === 'raised') {
+                if (link.startsWith('#') || link === '') {
                     return (
-                        <button className='button' onClick={onClick}>
+                        <button className='button' onClick={onClick} type={type}>
                             <a className="button-raised" href={link}>
                                 <span>{text}</span>
                             </a>
@@ -99,7 +99,7 @@ export default function Button({
                     )
                 } else {
                     return (
-                        <button className='button' onClick={onClick}>
+                        <button className='button' onClick={onClick} type={type}>
                             <Link className="button-raised" to={link}>
                                 <span>{text}</span>
                             </Link>
@@ -108,13 +108,13 @@ export default function Button({
                 }
                 //Returns null for no rendering of the button
             } else {
-                console.error('Unknown button type or missing icon for icon-type Button!');
+                console.error('Unknown button type!');
                 return null;
             }
         }
         //Returns null for no rendering of the button
     } else {
-        console.error(`Button's link or type should be strings only for Button component`);
+        console.error(`Link and Variant are required and must be strings only for Button component`);
         return null;
     }
 }
