--- frontend/src/Components/LoginForm/LoginForm.css
@@ -1,156 +1,0 @@
-:root {
-  --primary: #F74A5C;
-  --black: #3E3E3E;
-  --gray: #8C8C8C;
-  --white: #FFFFFF;
-  --blue: #4285EB;
-}
-
-.form-container {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-
-a {
-  text-decoration: none;
-  color: var(--primary);
-  cursor: pointer !important;
-}
-
-.input-group {
-  position: relative;
-}
-
-input[type="text"],
-input[type="email"],
-input[type="password"],
-input[type="number"] {
-  width: 100%;
-  height: 40px;
-  line-height: 40px;
-  border: 1px solid var(--gray);
-  border-radius: 0.25rem;
-  padding: 0 1rem;
-  margin-bottom: 1rem;
-  box-shadow: 0px 5px 15px -5px rgb(0 0 0 / 15%);
-  transition: all .2s ease-in-out;
-  outline: none;
-}
-
-input[type="text"]:focus,
-input[type="email"]:focus,
-input[type="password"]:focus,
-input[type=number] {
-  border-color: var(--blue);
-  box-shadow: 0px 5px 15px -5px rgba(66, 133, 235, 0.5);
-}
-
-input[type=number]::-webkit-inner-spin-button,
-input[type=number]::-webkit-outer-spin-button {
--webkit-appearance: none;
-margin: 0;
-}
-
-input[type=number] { -moz-appearance:textfield; }
-
-.text-primary {
-  color: var(--primary);
-}
-
-.text-center {
-  text-align: center;
-}
-
-.btn {
-  height: 40px;
-  width: 100%;
-  border: 1px solid var(--gray);
-  color: var(--black);
-  margin: 1rem 0;
-  padding: 0 1rem;
-  border-radius: 20px;
-  transition: all .2s ease-in-out;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-
-.btn.btn-primary {
-  background: var(--primary);
-  color: var(--white);
-  border-color: var(--primary);
-}
-
-.btn.btn-primary:hover {
-  cursor: pointer;
-}
-
-.btn.btn-google {
-  background: var(--white);
-  border-color: var(--gray);
-  cursor: pointer;
-}
-
-.btn.btn-facebook {
-  background: var(--blue);
-  color: var(--white);
-  border-color: var(--blue);
-}
-
-.login-wrapper {
-  margin: 1rem 0;
-  padding: 1rem 2rem;
-  background: linear-gradient(130deg, rgb(255 255 255 / 75%), transparent);
-  border-radius: 1rem;
-  backdrop-filter: blur(12px);
-  max-width: 380px;
-}
-
-.btn > .iconify {
-  margin-right: 0.5rem;
-}
-
-.password-button {
-  position: absolute;
-  height: 40px;
-  right: 0.5rem;
-  display: flex;
-  top: 0;
-  align-items: center;
-  font-size: 1.3rem;
-  color: var(--gray);
-  border: none;
-  background-color: transparent;
-  cursor: pointer;
-}
-
-.password-button:before {
-  content: '';
-  width: 2px;
-  height: 20px;
-  background: var(--gray);
-  position: absolute;
-  transform: rotate(45deg);
-  left: 10px;
-  transition: all .2s ease-in-out;
-}
-
-.password-button.active:before {
-  transform: rotate(45deg) scale(0);
-}
-
-.divider {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  margin: 1rem 0;
-}
-
-.divider:after,
-.divider:before {
-  content: '';
-  width: 45%;
-  height: 1px;
-  background: var(--gray);
-}
