+++ frontend/src/Pages/CourseDetail/CourseDetail.js
@@ -16,8 +16,7 @@ import withReactContent from "sweetalert2-react-content";
 import ReactPlayer from 'react-player';
 import CardsVideos from "../../Components/CardsVideos/CardsVideos";
 
-const video= 'https://www.youtube.com/watch?v=QrDJ9zv0Pwg&ab_channel=ENTERTAIMENTNOW'
-
+const video = 'https://www.youtube.com/watch?v=QrDJ9zv0Pwg&ab_channel=ENTERTAIMENTNOW'
 
 export default function CourseDetail({ isLoggedIn }) {
 	const { id } = useParams();
@@ -157,7 +156,7 @@ export default function CourseDetail({ isLoggedIn }) {
 				{courseDetail && (
 					<div className="course-details_back-btn">
 						<Button
-							type={"raised"}
+							btnVariant={"raised"}
 							text={"Volver a cursos"}
 							link={"/home"}
 						></Button>
@@ -205,18 +204,19 @@ export default function CourseDetail({ isLoggedIn }) {
 							<p>{courseDetail?.description}</p>
 							<h2>$ {courseDetail?.price}</h2>
 							<div className="actionsButtons">
-								{isLoggedIn === "teacher" &&
-								userCredentials.id === courseDetail.teacherID ? null : (
-									<div className="buyBtn">
-										<Button
-											icon={"bi:cart-plus"}
-											type={"raised-icon"}
-											text={"Agregar al carrito"}
-											onClick={handleAddCart}
-											link={""}
-										></Button>
-									</div>
-								)}
+								{(isLoggedIn === 'teacher' && userCredentials.id === courseDetail.teacherID)
+									? null
+									: (
+										<div className="buyBtn">
+											<Button
+												icon={"bi:cart-plus"}
+												btnVariant={"raised-icon"}
+												text={"Agregar al carrito"}
+												onClick={handleAddCart}
+												link={""}
+											></Button>
+										</div>
+									)}
 							</div>
 						</div>
 					</main>
@@ -224,8 +224,7 @@ export default function CourseDetail({ isLoggedIn }) {
 					<Loader />
 				)}
 			</div>
-			<CardsVideos id={id}/>
+			<CardsVideos id={id} />
 		</section >
-
 	);
 }
