+++ frontend/src/Pages/UserHome/UserHome.js
@@ -22,7 +22,7 @@ export default function UserHome({ isLoggedIn }) {
                 {/* <Button
                     link={''}
                     text={'Mostrar recientes'}
-                    type='flat'
+                    btnVariant='flat'
                     onClick={() => {
 
                         setHasRecents(hasRecents => !hasRecents)
