+++ frontend/src/index.css
@@ -39,6 +39,49 @@ html {
   margin   : 0 auto;
 }
 
+button,
+a {
+  text-decoration: none;
+  cursor         : pointer !important;
+}
+
+input[type=text],
+input[type=password],
+input[type=email] {
+  height          : 80%;
+  width           : 100%;
+  padding         : 1rem;
+  border-radius   : .6rem;
+  font-size       : 2ch;
+  font-weight     : 600;
+  text-align      : left;
+  background-color: var(--ash-gray-lighter);
+  color           : var(--nice-black);
+  transition      : all .2s ease-in-out;
+  border          : none;
+  outline         : .2rem solid var(--ash-gray);
+}
+
+input[type=text]:hover:not([disabled]),
+input[type=password]:hover:not([disabled]),
+input[type=email]:hover:not([disabled]),
+input[type=text]:focus-visible:not([disabled]),
+input[type=password]:focus-visible:not([disabled]),
+input[type=email]:focus-visible:not([disabled]) {
+  outline            : .3rem solid var(--lighter-purple);
+  /* background-color: whitesmoke; */
+  box-shadow         : 0px 0.4rem 1rem -0.2rem var(--lighter-purple);
+}
+
+input[type=text]:disabled,
+input[type=password]:disabled,
+input[type=email]:disabled {
+  background-color: hsl(150, 9%, 87%);
+  color           : hsl(0, 0%, 54%);
+  border          : none;
+  outline         : .2rem solid var(--ash-gray);
+}
+
 .dark {
   filter: invert(1);
 }
